requirements.txt
requests
beautifulsoup4
pandas
gtts
python-dotenv
